rename and update imports to CTALink
