+++ src/components/Cards/StandardCard/index.js
@@ -6,7 +6,7 @@ import { cards, color, fontSize, mixins, spacing, withThemes } from '../../../st
 import { StandardUserAttributions } from '../shared/UserAttributions/UserAttributions';
 import Attributions from '../shared/Attributions';
 import Badge from '../../Badge/Badge';
-import CtaLink from '../shared/CtaLink';
+import CtaLink from '../shared/CtaLink/CtaLink';
 import FavoriteButton from '../shared/FavoriteButton';
 import Image from '../shared/Image';
 import ImageCollage from '../shared/ImageCollage';
